pyproject.toml
[project]
name = "flask-configurator"
version = "0.1.2"
description = "Env and file based configuration for Flask"
authors = [
    {"name" = "Maxime Bouroumeau-Fuseau", email = "maxime.bouroumeau@gmail.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "flask~=3.0",
    "python-dotenv~=1.0",
    "pyyaml~=6.0",
    "watchdog~=6.0",
]

[project.urls]
Repository = "https://github.com/hyperflask/flask-configurator"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "flask_configurator"
module-root = ""
source-exclude = ["example"]